oop update dwind phys
